+++ src/main/java/com/fulvo/backend/services/TournamentService.java
@@ -6,13 +6,17 @@ import com.fulvo.backend.dto.match.FixtureResponse;
 import com.fulvo.backend.dto.match.MatchResponse;
 import com.fulvo.backend.dto.team.TeamTournamentRequest;
 import com.fulvo.backend.dto.tournament.TournamentRequest;
+import com.fulvo.backend.dto.tournament.TournamentResponse;
 import com.fulvo.backend.models.*;
 import com.fulvo.backend.repositories.TournamentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -25,7 +29,7 @@ public class TournamentService {
     private final TournamentTeamHelperService tournamentTeamHelperService;
     private final MatchService matchService;
 
-    public GenericResponse createTournament(TournamentRequest request) {
+    public TournamentResponse createTournament(TournamentRequest request) {
         User admin = userService.getUser();
 
         boolean exist = tournamentRepository.existsByNameAndAdmin(request.getName(), admin);
@@ -34,19 +38,33 @@ public class TournamentService {
 
         Tournament tournament = Tournament.builder()
                 .name(request.getName())
+                .teamsMax(request.getTeamsMax())
+                .teamsMin(request.getTeamsMin())
+                .roundTrip(request.getRoundTrip())
+                .teams(0)
                 .admin(admin)
                 .build();
 
         tournamentRepository.save(tournament);
-        return GenericResponse.builder()
+        return TournamentResponse.builder()
                 .name(tournament.getName())
+                .teamsMax(tournament.getTeamsMax())
+                .teamsMin(tournament.getTeamsMin())
                 .build();
     }
 
     public GenericResponse deleteTournament(TournamentRequest request) {
         User admin = userService.getUser();
-        Tournament tournament = tournamentRepository.findByNameAndAdmin(request.getName(), admin)
+        Tournament tournament = tournamentRepository.findById(request.getId())
                 .orElseThrow(() -> new RuntimeException("No se encontro el torneo"));
+        if (admin != tournament.getAdmin())
+            throw new RuntimeException("No sos el administrador del torneo");
+
+        List<Match> matchList = matchService.findAllByTournament(tournament);
+        if (!matchList.isEmpty()){
+            matchService.deleteAll(matchList);
+        }
+
         List <Scoreboard> scoreboardList = scoreboardService.findAllByTournament(tournament);
         if (!scoreboardList.isEmpty()){
             scoreboardService.deleteAll(scoreboardList);
@@ -68,16 +86,25 @@ public class TournamentService {
 
     public GenericResponse startTournament(TournamentRequest request) {
         User admin = userService.getUser();
-        Tournament tournament = tournamentRepository.findByNameAndAdmin(request.getName(), admin)
+        Tournament tournament = tournamentRepository.findById(request.getId())
                 .orElseThrow(() -> new RuntimeException("No se encontro el torneo"));
-        List<Scoreboard> scoreboardList = scoreboardService.findAllByTournament(tournament);
-        List<Team> teams = tournamentTeamHelperService.getAllTeams(scoreboardList);
+        if (admin != tournament.getAdmin())
+            throw new RuntimeException("No sos el administrador del torneo");
+
+        List<Scoreboard> teams = scoreboardService.findAllByTournament(tournament);
+        if (teams.size() < tournament.getTeamsMin())
+            throw new RuntimeException("No se alcanzó la cantidad mínima de equipos");
+
+        if (teams.size() % 2 != 0){
+            scoreboardService.joinTournament(tournamentTeamHelperService.getTeam(0),tournament);
+            teams.add(scoreboardService
+                    .findByTeamAndTournament(tournamentTeamHelperService.getTeam(0), tournament));
+        }
 
         Collections.shuffle(teams);
-        matchService.generateMatches(teams);
+        matchService.generateMatches(teams, !tournament.getRoundTrip());
         if (tournament.getRoundTrip()){
-            teams.reversed();
-            matchService.generateMatches(teams);
+            matchService.generateMatches(teams, tournament.getRoundTrip());
         }
 
         return GenericResponse.builder()
@@ -90,15 +117,44 @@ public class TournamentService {
     public FixtureResponse getFixture(TournamentRequest request) {
         Tournament tournament = tournamentRepository.findById(request.getId())
                 .orElseThrow(() -> new RuntimeException("No se encontro el torneo"));
-        return FixtureResponse.builder().build();
+        int dates = tournament.getRoundTrip() ? (tournament.getTeams()-1) * 2 : tournament.getTeams()-1;
+        List<DateResponse> matches = new ArrayList<>();
+        for(int i=1; i<=dates; i++){
+            matches.add(getDateResponse(tournament, i));
+        }
+
+        return FixtureResponse.builder()
+                .tournament(tournament.getName())
+                .matches(matches)
+                .build();
     }
+
 
     public DateResponse getFixture(TournamentRequest request, Integer date) {
-        return DateResponse.builder().build();
+        Tournament tournament = tournamentRepository.findById(request.getId())
+                .orElseThrow(() -> new RuntimeException("No se encontro el torneo"));
+
+        return getDateResponse(tournament, date);
     }
 
-    public MatchResponse getMatch(){
-        return MatchResponse.builder().build();
+    private DateResponse getDateResponse(Tournament tournament, int date){
+        List<Match> matches = matchService.getMatchesByDate(tournament, date);
+        List<MatchResponse> matchResponses = matches.stream()
+                .map(this::getMatchResponse)
+                .collect(Collectors.toList());
+        return DateResponse.builder()
+                .date(date)
+                .matches(matchResponses)
+                .build();
+    }
+
+    public MatchResponse getMatchResponse(Match match){
+        return MatchResponse.builder()
+                .homeTeam(match.getHomeTeam().getTeam().getName())
+                .awayTeam(match.getAwayTeam().getTeam().getName())
+                .day(match.getDay())
+                .date(match.getDate())
+                .build();
     }
 
 }
